// api/src/modules/position/constants/enums/errors.enum.ts
export enum ErrorMessage {
  POSITION_ID_EMPTY_ERROR = 'Bạn vui lòng chọn [chứ vụ].',

  ID_NAN_ERROR = 'Giá trị [id] không hợp lệ.',

  NO_CONTENT = 'Không có dữ liệu hiển thị.',

  OPERATOR_POSITION_ERROR = 'Lưu thông tin [chức vụ] thất bại.',
  POSITION_NOT_FOUND_ERROR = '[Chức vụ] không tồn tại (id: %s)',
}